Add files via upload
This dataset was downloaded from Kaggle and it contains a set of informations about California housing prices. I predict and test my model by using the location to predict the price of the houses in California. It was observed that the majority of the houses that were closed to the coast had higher prices when compared to other locations in the state.